fix(OUT-746): company unassignment logic

+++ src/app/api/webhook/webhook.service.ts
@@ -10,8 +10,15 @@ import Bottleneck from 'bottleneck'
 import { getInProductNotificationDetails } from '@api/notification/notification.helpers'
 import { NotificationTaskActions } from '@api/core/types/tasks'
 import { NotificationService } from '@api/notification/notification.service'
+import User from '@api/core/models/User.model'
 
 class WebhookService extends BaseService {
+  private copilot
+  constructor(user: User, customCopilotApiKey?: string) {
+    super(user, customCopilotApiKey)
+    this.copilot = new CopilotAPI(this.user.token)
+  }
+
   async parseWebhook(req: NextRequest): Promise<WebhookEvent> {
     const webhookEvent = WebhookSchema.safeParse(await req.json())
     if (!webhookEvent.success) {
@@ -48,9 +55,8 @@ class WebhookService extends BaseService {
 
   async handleClientCreated(assigneeId: string) {
     // First fetch all the current non-complete tasks for client's company, if exists
-    const copilot = new CopilotAPI(this.user.token)
-    const client = await copilot.getClient(assigneeId)
-    const company = await copilot.getCompany(client.companyId)
+    const client = await this.copilot.getClient(assigneeId)
+    const company = await this.copilot.getCompany(client.companyId)
     if (company.name === '') {
       // Client does not have a fixed company!
       return
@@ -59,12 +65,16 @@ class WebhookService extends BaseService {
     const tasksService = new TasksService(this.user)
     const tasks = await tasksService.getIncompleteTasksForCompany(company.id)
     if (!tasks.length) return
+    console.log('creating task notifications for', client.givenName, tasks.length)
 
     // Then trigger appropriate notifications
     const bottleneck = new Bottleneck({ minTime: 250, maxConcurrent: 2 })
     const createNotificationPromises = []
 
-    const internalUsers = (await copilot.getInternalUsers()).data
+    const internalUsers = (await this.copilot.getInternalUsers({ limit: 10_000 })).data
+    const existingNotifications = await this.db.clientNotification.findMany({
+      where: { clientId: assigneeId },
+    })
 
     for (let task of tasks) {
       const actionUser = internalUsers.find((iu) => iu.id === task.createdById)
@@ -74,6 +84,9 @@ class WebhookService extends BaseService {
           `webhookController :: could not get action user for company task ${task.id}`,
         )
       }
+
+      const existingNotification = existingNotifications.find((notification) => notification.taskId === task.id)
+      if (existingNotification) continue
 
       const actionUserName = `${actionUser.givenName} ${actionUser.familyName}`
       const inProduct = getInProductNotificationDetails(actionUserName, task, company.name)[
@@ -85,9 +98,10 @@ class WebhookService extends BaseService {
         // If any of the given action is not present in details obj, that type of notification is not sent
         deliveryTargets: { inProduct },
       }
-      createNotificationPromises.push(bottleneck.schedule(() => copilot.createNotification(notificationDetails)))
+      createNotificationPromises.push(bottleneck.schedule(() => this.copilot.createNotification(notificationDetails)))
     }
     const notifications = await Promise.all(createNotificationPromises)
+    console.log('create notifications length', notifications.length)
 
     // Now add appropriate records to our ClientNotifications table
     const insertBottleneck = new Bottleneck({ minTime: 100, maxConcurrent: 4 })
@@ -99,18 +113,21 @@ class WebhookService extends BaseService {
         insertBottleneck.schedule(() => notificationService.addToClientNotifications(tasks[i], notifications[i])),
       )
     }
-    await Promise.all(insertPromises)
+    const inserts = await Promise.all(insertPromises)
+    console.log('inserts length', inserts.length)
   }
 
   async handleUserDeleted(assigneeId: string, assigneeType: AssigneeType) {
+    if (assigneeType === AssigneeType.company) return
+
     const tasksService = new TasksService(this.user)
     // Delete corresponding tasks
     console.info(`Deleting all tasks for ${assigneeType} ${assigneeId}`)
     await tasksService.deleteAllAssigneeTasks(assigneeId, assigneeType)
 
-    // Delete corresponding notifications
-    const notificationService = new NotificationService(this.user)
-    await notificationService.readAllUserNotifications(assigneeId, assigneeType)
+    // So how do we delete corresponding notifications, you ask? Good question.
+    // Copilot calls multiple `client.updated` when unassigning clients companies after `company.deleted` is called.
+    // So handling deleting company task notifications should be done in `client.updated` company unassignment logic instead
   }
 
   async handleClientUpdated(data: ClientUpdatedEventData) {
@@ -119,54 +136,78 @@ class WebhookService extends BaseService {
       companyId: newCompanyId,
       previousAttributes: { companyId: prevCompanyId },
     } = data
+    console.log('changed company', newCompanyId, prevCompanyId)
     // If company hasn't been changed - don't bother with any of this
-    if (prevCompanyId === newCompanyId) return
+    if (prevCompanyId === newCompanyId || !prevCompanyId) return
 
     // Only delete prev notifications if client had a valid company before
-    const copilot = new CopilotAPI(this.user.token)
-    const prevCompany = prevCompanyId ? await copilot.getCompany(prevCompanyId) : null
+    const prevCompany = await this.copilot.getCompany(prevCompanyId)
+    const newCompany = await this.copilot.getCompany(newCompanyId)
+
+    if (prevCompany.name === '' && newCompany.name !== '') {
+      await this.handleCompanyAssignment(clientId)
+    } else if (newCompany.name === '' && prevCompany.name !== '') {
+      await this.handleCompanyUnassignment(clientId, prevCompanyId)
+    } else if (prevCompany.name !== '' && newCompany.name !== '') {
+      await this.handleCompanyUnassignment(clientId, prevCompanyId)
+      await this.handleCompanyAssignment(clientId)
+    } else {
+      // This should never happen, still we want to be reported if this ever does
+      throw new APIError(httpStatus.INTERNAL_SERVER_ERROR, 'Could not identify company change')
+    }
+  }
+
+  private async handleCompanyAssignment(clientId: string) {
+    // Notifications logic is the same as when a new client is created
+    await this.handleClientCreated(clientId)
+  }
+
+  private async handleCompanyUnassignment(clientId: string, prevCompanyId: string) {
+    const company = await this.copilot.getCompany(prevCompanyId)
+    console.log('unassigned', company)
+
     // NOTE: If prev company was not a valid company, instead a randomly generated placeholder company,
     // prevCompany.name will be an empty string
-    if (prevCompany && prevCompany.name !== '') {
-      // First find all tasks related to previous company
-      const prevCompanyTasks = await this.db.task.findMany({
-        where: {
-          assigneeId: prevCompanyId,
-          assigneeType: AssigneeType.company,
-          workflowState: {
-            type: { not: StateType.completed },
-          },
-        },
-      })
-      const prevCompanyTaskIds = prevCompanyTasks.map((task) => task.id)
-
-      // Find all triggered notifications for this client, on behalf of prev company
-      const prevCompanyNotifications = await this.db.clientNotification.findMany({
-        where: {
-          taskId: { in: prevCompanyTaskIds },
-          clientId,
-        },
-      })
-      const notificationIds = prevCompanyNotifications.map((notification) => notification.id)
-
-      // Delete all task notifications triggered for client for previous company
-      const deletePromises = []
-      const bottleneck = new Bottleneck({ minTime: 250, maxConcurrent: 2 })
-
-      for (let notification of prevCompanyNotifications) {
-        deletePromises.push(
-          bottleneck.schedule(() => {
-            return copilot.markNotificationAsRead(notification.notificationId)
-          }),
-        )
-      }
-      await Promise.all(deletePromises)
-      await this.db.clientNotification.deleteMany({ where: { id: { in: notificationIds } } })
+    if (!company.name) {
+      throw new APIError(httpStatus.INTERNAL_SERVER_ERROR, 'Cannot unassign from a company that does not exist')
     }
 
-    // Trigger new notifications for new company's tasks (if exists)
-    if (!newCompanyId) return
-    await this.handleClientCreated(clientId)
+    // First find all tasks related to previous company
+    const prevCompanyTasks = await this.db.task.findMany({
+      where: {
+        assigneeId: company.id,
+        assigneeType: AssigneeType.company,
+        workflowState: {
+          type: { not: StateType.completed },
+        },
+      },
+    })
+    const prevCompanyTaskIds = prevCompanyTasks.map((task) => task.id)
+    console.log('unassigned tasks', prevCompanyTaskIds.length)
+
+    // Find all triggered notifications for this client, on behalf of prev company
+    const prevCompanyNotifications = await this.db.clientNotification.findMany({
+      where: {
+        taskId: { in: prevCompanyTaskIds },
+        clientId,
+      },
+    })
+    const notificationIds = prevCompanyNotifications.map((notification) => notification.id)
+    console.log('unassigned nots', notificationIds.length)
+
+    // Delete all task notifications triggered for client for previous company
+    const deletePromises = []
+    const bottleneck = new Bottleneck({ minTime: 250, maxConcurrent: 2 })
+
+    for (let notification of prevCompanyNotifications) {
+      deletePromises.push(
+        bottleneck.schedule(() => {
+          return this.copilot.markNotificationAsRead(notification.notificationId)
+        }),
+      )
+    }
+    await Promise.all(deletePromises)
+    await this.db.clientNotification.deleteMany({ where: { id: { in: notificationIds } } })
   }
 }
